refactor: move every thing to its own file
