+++ pkg/rbac/permissions/gateway_policies.go
@@ -2,7 +2,7 @@ package permissions
 
 import "github.com/unkeyed/unkey/pkg/urn"
 
-// ReadPolicy authorizes reading gateway policy resources.
+// ReadPolicy authorizes reading an environment's gateway policies.
 //
 // Valid resource: urn.GatewayPolicy.
 type ReadPolicy struct{}
@@ -10,10 +10,12 @@ type ReadPolicy struct{}
 func (ReadPolicy) ActionFor(urn.GatewayPolicy) {}
 func (ReadPolicy) String() string              { return "read_policy" }
 
-// WritePolicy authorizes creating, replacing, updating, or deleting gateway
-// policy resources.
+// WritePolicy authorizes replacing an environment's entire gateway policy list
+// or updating a single policy in place.
 //
-// Valid resource: urn.GatewayPolicy.
+// Valid resource: urn.GatewayPolicy. Replacing the list uses a wildcard policy
+// id because it covers policies that do not exist yet when the request is
+// authorized.
 type WritePolicy struct{}
 
 func (WritePolicy) ActionFor(urn.GatewayPolicy) {}
