+++ pkg/urn/urn_test.go
@@ -233,8 +233,6 @@ func TestResourceCatalogHelpers(t *testing.T) {
 	require.Equal(t, "unkey:v1:ws_123:projects/proj_123/apps/app_123/environments/env_123/deployments/dep_123/**", workspace.Project("proj_123").App("app_123").Environment("env_123").Deployment("dep_123").Any().String())
 	require.Equal(t, "unkey:v1:ws_123:projects/proj_123/apps/app_123/environments/env_123/domains/dom_123", workspace.Project("proj_123").App("app_123").Environment("env_123").Domain("dom_123").String())
 	require.Equal(t, "unkey:v1:ws_123:projects/proj_123/apps/app_123/environments/env_123/variables/var_123", workspace.Project("proj_123").App("app_123").Environment("env_123").Variable("var_123").String())
-	require.Equal(t, "unkey:v1:ws_123:projects/proj_123/apps/app_123/environments/env_123/gateway", workspace.Project("proj_123").App("app_123").Environment("env_123").Gateway().String())
-	require.Equal(t, "unkey:v1:ws_123:projects/proj_123/apps/app_123/environments/env_123/gateway/**", workspace.Project("proj_123").App("app_123").Environment("env_123").Gateway().Any().String())
 	require.Equal(t, "unkey:v1:ws_123:projects/proj_123/apps/app_123/environments/env_123/gateway/policies/pol_123", workspace.Project("proj_123").App("app_123").Environment("env_123").Gateway().Policy("pol_123").String())
 	require.Equal(t, "unkey:v1:ws_123:portals/portal_123/sessions/session_123", workspace.Portal("portal_123").Session("session_123").String())
 	require.Equal(t, "unkey:v1:ws_123:portals/portal_123", workspace.Portal("portal_123").String())
